requirements.txt
numpy==1.19.2
pandas==1.3.5
imutils==0.5.3
pillow==8.4.0
requests==2.27.1
opencv-python-headless==4.5.1.48
matplotlib==3.5.1
tensorflow==2.4.1
scikit-learn==0.22.2